refactor: adaptor 계층 내 controller 수정
