--- src/main/java/online/bottler/complaint/application/dto/request/ComplaintRequestDTO.java
@@ -1,9 +1,0 @@
-package online.bottler.complaint.application.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record ComplaintRequestDTO(
-        @NotBlank(message = "신고 사유를 작성해주세요.")
-        String description
-) {
-}
