+++ src/main/java/online/bottler/notification/adaptor/in/web/UnsubscriptionRequest.java
@@ -1,0 +1,6 @@
+package online.bottler.notification.adaptor.in.web;
+
+public record UnsubscriptionRequest(
+        String token
+) {
+}
