--- src/main/java/online/bottler/notification/application/dto/request/SubscriptionRequestDTO.java
@@ -1,6 +1,0 @@
-package online.bottler.notification.application.dto.request;
-
-public record SubscriptionRequestDTO(
-        String token
-) {
-}
