--- src/main/java/online/bottler/notification/application/dto/request/UnsubscriptionRequestDTO.java
@@ -1,6 +1,0 @@
-package online.bottler.notification.application.dto.request;
-
-public record UnsubscriptionRequestDTO(
-        String token
-) {
-}
